Solution for day 4.
Not too happy with how this turned out. Looks pretty inefficient /
overly complicated.

Started out putting way too many checks on the input data, resulting
in too low numbers of passports for part 1, as the assignment was only
to look for records with the necessary keys present and no further
validation.